requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
cycler==0.10.0
Django==3.2.7
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzlocal==3.0
